fix(angular): fix 6 compiler divergences from Angular reference
1. viewport({trigger: ...}) now keeps empty options object (was dropping it)
2. Timer delays preserve fractional precision (u32 -> f64, e.g. 1.5ms)
3. Pipe detection in @for track now traverses all AST expression types
4. hover/interaction triggers enforce max 1 parameter; never removed from on-triggers
5. @for let assignment a=b=c correctly rejected (split instead of splitn)
6. @for expression parse failure no longer emits extra missing-track error

+++ crates/oxc_angular_compiler/src/ir/ops.rs
@@ -1062,7 +1062,7 @@ pub struct DeferOnOp<'a> {
     /// Target name for local ref targeting.
     pub target_name: Option<Atom<'a>>,
     /// Timer delay.
-    pub delay: Option<u32>,
+    pub delay: Option<f64>,
     /// Viewport options (for viewport trigger).
     pub options: Option<Box<'a, IrExpression<'a>>>,
 }
